memory opt

+++ tests/BruteForce.cpp
@@ -2,8 +2,8 @@
 
 BruteForce::BruteForce(const std::string &inputString) : string(inputString) {}
 
-std::vector<int> BruteForce::search(const std::string &pattern) const {
-    std::vector<int> matchIndexes;
+std::vector<size_t> BruteForce::search(const std::string &pattern) const {
+    std::vector<size_t> matchIndexes;
 
     for (size_t i = 0; i + pattern.length() <= this->string.length(); ++i) {
         size_t j = 0;
